# sources/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/users": {
      "post": {
        "operationId": "UsersController_createUser",
        "summary": "Регистрация пользователя",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь зарегистрирован и пользователю пришло на почту ссылка подтверждения.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Логин занят другим пользователем. \n\nE-mail занят другим пользователем. \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзация не совершена, так как не удалось отправить письмо с ссылкой активации на почту. \n\nЧто-то пошло не так на сервере. \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_users",
          "any"
        ]
      },
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Получить информацию о профиле",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "default": {
            "description": "Получили информацию о пользователе",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_users",
          "any"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/users/activate-account/{token}": {
      "get": {
        "operationId": "UsersController_activateAccount",
        "summary": "Активация аккаунта",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь подтвердил зарегистрированный аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Ссылка не действительна, так как прошло 24 часа (токен просрочен). \n\nСсылка не действительна, так как токен поддельный (токен не прошел проверку валидности). \n\nСсылка не действительна, так как токен не зарегстрирован в БД (токен прошел проверку, но его нет в БД). \n\nСсылка не действительна, так как аккаунт был активирован. \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Не совершено обновление статуса пользователя. \n\nЧто-то пошло не так на сервере. \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_users",
          "any"
        ]
      }
    },
    "/api/v1/users/change-email": {
      "patch": {
        "operationId": "UsersController_updateEmail",
        "summary": "Запрос на смену электронной почты на новую электронную почту",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Заявка отправлена на новую почту и предупреждение отправлено на старую почту.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Новая почта совпадает с текущей.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Cлишком много запросов о смене электроной почты за не прошедшие 3 часа.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзация не совершена, так как нет таблицы в БД (не сохранения запись о смене старой электронной почты на новую). \n\nТранзация не совершена, так как не отправлено письмо на новую и старую почту. \n\nЧто-то пошло не так на сервере. \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_users",
          "user"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/users/change-email/{token}/confirm": {
      "get": {
        "operationId": "UsersController_confirmChangeEmail",
        "summary": "Подтверждаем смену на новую электронную почту",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Электронная почта изменена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Ссылка не действительна, так как прошло 3 часа (токен просрочен). \n\nСсылка не действительна, так как токен подделан (токен не прошел валидацию). \n\nСсылка не действительна, так как токена не зарегистрирован в БД. \n\nСсылка не действительна, так как почта была сменена, либо заявка на смену почты была отклонена. \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзация не совернеша (заявка не отмечена закрытой, пользователю не поменял email на новый). \n\nЧто-то пошло не так на сервере. \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_users",
          "any"
        ]
      }
    },
    "/api/v1/users/change-email/{token}/delete": {
      "get": {
        "operationId": "UsersController_deleteChangeEmail",
        "summary": "Отменяем смену на новую электронную почту",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Заявка на смену электронной почты отменена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Cсылка не действительна, так как прошло 3 часа (токен просрочен). \n\nCсылка не действительна, так как токен подделан (токен не прошел валидацию). \n\nСсылка не действительна, так как токен не зарегистрирован в БД. \n\nСсылка не действительна, так как почта была сменена, либо заявка на смену почты была отклонена. \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_users",
          "any"
        ]
      }
    },
    "/api/v1/users/forget-password": {
      "post": {
        "operationId": "UsersController_forgetPassword",
        "summary": "Забыли пароль",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "На электронную почту отправлен логин и новый сгенерированный пароль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Пользователя с такой электронной почтой не существует",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзация не совернеша, так как сгенерированный пароль не записался в БД. \n\nТранзация не совернеша, так как письмо с логином и паролем не отправлено на электронную почту. \n\nЧто-то пошло не так на сервере. \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_users",
          "any"
        ]
      }
    },
    "/api/v1/users/change-password": {
      "patch": {
        "operationId": "UsersController_updatePassword",
        "summary": "Меняем пароль",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пароль изменен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Не тот старый пароль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзация не совернеша, так как новый пароль не записан в БД. \n\nТранзация не совернеша, так как не удалось завершить все сессии. \n\nЧто-то пошло не так на сервере. \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_users",
          "user"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "REST API документация",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "access token - токен доступа",
        "in": "header"
      },
      "refresh-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "refresh token - токен обновления",
        "in": "header"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "dp_login": {
            "type": "string",
            "example": "user123"
          },
          "dp_email": {
            "type": "string",
            "example": "user123@example.com"
          },
          "dp_password": {
            "type": "string",
            "example": "secret123"
          },
          "dp_unp": {
            "type": "string",
            "example": "1234567890123"
          },
          "dp_nameLegalEntity": {
            "type": "string",
            "example": "Частное торгово-производственное унитарное предприятие \"Иванов\""
          },
          "dp_shortNameLegalEntity": {
            "type": "string",
            "example": "Частное предприятие \"Иванов\""
          },
          "dp_address": {
            "type": "string",
            "example": "224017, Беларусь, Брестская обл., г. Брест, ул. Московская, д. 267"
          },
          "dp_receptionPhone": {
            "type": "string",
            "example": "+375331112233"
          },
          "dp_firstName": {
            "type": "string",
            "example": "Иван"
          },
          "dp_lastName": {
            "type": "string",
            "example": "Иванович"
          },
          "dp_middleName": {
            "type": "string",
            "example": "Иванов"
          }
        },
        "required": [
          "dp_login",
          "dp_email",
          "dp_password",
          "dp_unp",
          "dp_nameLegalEntity",
          "dp_shortNameLegalEntity",
          "dp_address",
          "dp_receptionPhone",
          "dp_firstName",
          "dp_lastName",
          "dp_middleName"
        ]
      },
      "CreateUserResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 201
          },
          "message": {
            "type": "string"
          },
          "dp_accessToken": {
            "type": "string",
            "example": "токен"
          },
          "dp_refreshToken": {
            "type": "string",
            "example": "токен"
          }
        },
        "required": [
          "statusCode",
          "message",
          "dp_accessToken",
          "dp_refreshToken"
        ]
      },
      "HttpResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "сообщение"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
      "GetUserDto": {
        "type": "object",
        "properties": {
          "dp_id": {
            "type": "number"
          },
          "dp_unp": {
            "type": "string"
          },
          "dp_nameLegalEntity": {
            "type": "string"
          },
          "dp_shortNameLegalEntity": {
            "type": "string"
          },
          "dp_address": {
            "type": "string"
          }
        },
        "required": [
          "dp_id",
          "dp_unp",
          "dp_nameLegalEntity",
          "dp_shortNameLegalEntity",
          "dp_address"
        ]
      },
      "ChangeEmailDto": {
        "type": "object",
        "properties": {
          "dp_email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "dp_email"
        ]
      },
      "ForgetPasswordDto": {
        "type": "object",
        "properties": {
          "emailOrLogin": {
            "type": "string",
            "example": "user123@example.com"
          }
        },
        "required": [
          "emailOrLogin"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "dp_oldPassword": {
            "type": "string",
            "example": "secret123"
          },
          "dp_newPassword": {
            "type": "string",
            "example": "secret321"
          }
        },
        "required": [
          "dp_oldPassword",
          "dp_newPassword"
        ]
      },
...
    }
  }
}